Refactor DatabaseService to use injected RealmConfiguration
- Add `RealmConfiguration` to `DatabaseService` primary constructor.
- Remove Realm initialization and configuration logic from `DatabaseService` init block.
- Update `Realm.getDefaultInstance()` calls to `Realm.getInstance(realmConfiguration)`.
- Create `@Provides` function in `DatabaseModule` to supply `RealmConfiguration` and inject it into `DatabaseService`.
- Update `DatabaseServiceTest` and `RealmUserTest` to construct `DatabaseService` with the required local test configuration.
