+++ app/src/androidTest/java/org/ole/planet/myplanet/model/RealmUserTest.kt
@@ -55,7 +55,7 @@ class RealmUserTest {
             override val default: CoroutineDispatcher = Dispatchers.Unconfined
             override val unconfined: CoroutineDispatcher = Dispatchers.Unconfined
         }
-        databaseService = DatabaseService(ApplicationProvider.getApplicationContext(), testDispatcherProvider)
+        databaseService = DatabaseService(ApplicationProvider.getApplicationContext(), testDispatcherProvider, realmConfiguration)
 
         val mockSettings = mockk<SharedPreferences>(relaxed = true)
         val mockSharedPrefManager = mockk<SharedPrefManager>(relaxed = true)
