+++ app/src/main/java/org/ole/planet/myplanet/di/DatabaseModule.kt
@@ -9,6 +9,8 @@ import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineDispatcher
 import org.ole.planet.myplanet.data.DatabaseService
+import io.realm.Realm
+import io.realm.RealmConfiguration
 import org.ole.planet.myplanet.di.RealmDispatcher
 import org.ole.planet.myplanet.utils.DispatcherProvider
 
@@ -18,8 +20,30 @@ object DatabaseModule {
 
     @Provides
     @Singleton
-    fun provideDatabaseService(@ApplicationContext context: Context, dispatcherProvider: DispatcherProvider): DatabaseService {
-        return DatabaseService(context, dispatcherProvider)
+    fun provideRealmConfiguration(@ApplicationContext context: Context): RealmConfiguration {
+        Realm.init(context)
+        val targetLogLevel = if (org.ole.planet.myplanet.BuildConfig.DEBUG) io.realm.log.LogLevel.DEBUG else io.realm.log.LogLevel.ERROR
+        if (io.realm.log.RealmLog.getLevel() != targetLogLevel) {
+            io.realm.log.RealmLog.setLevel(targetLogLevel)
+        }
+        val currentConfig = Realm.getDefaultConfiguration()
+        return if (currentConfig == null || currentConfig.realmDirectory.name == Realm.DEFAULT_REALM_NAME) {
+            val config = RealmConfiguration.Builder()
+                .name(Realm.DEFAULT_REALM_NAME)
+                .schemaVersion(13)
+                .migration(org.ole.planet.myplanet.data.RealmMigrations())
+                .build()
+            Realm.setDefaultConfiguration(config)
+            config
+        } else {
+            currentConfig
+        }
+    }
+
+    @Provides
+    @Singleton
+    fun provideDatabaseService(@ApplicationContext context: Context, dispatcherProvider: DispatcherProvider, realmConfiguration: RealmConfiguration): DatabaseService {
+        return DatabaseService(context, dispatcherProvider, realmConfiguration)
     }
 
     @Provides
@@ -28,6 +52,4 @@ object DatabaseModule {
     fun provideRealmDispatcher(provider: RealmDispatcherProvider): CoroutineDispatcher {
         return provider
     }
-
-    // Realm initialization is handled in DatabaseService
 }
